fix some logic/id issues
